Create Gameboard factory with ship placement, attack handling, and sink detection
Implemented Gameboard factory to allow placement of ships at specific coordinates using the ship factory function. Added a receiveAttack function to determine if the attack hit a ship and send the ‘hit’ function to the correct ship, or record the coordinates of the missed shots. Modiefied Ship factory function to meet the functionality of the game. Also, included functionality to keep track of missed attacks and report if all ships have been sunk.