Create gameLoopFactory function
This commit adds a new function called gameLoopFactory which creates a game loop where the player and the computer take turns in attacking each other's ships. It uses functions from the domManipulatorFactory, playerFactory, and computerFactory to manipulate the DOM, create player and computer objects, and perform their respective actions during the game. The function also includes a startGame function that starts the game and a resetGame function that reloads the page. This commit also add the function to gameBoard factory to place ships randomly.